2018-12-08/src/day8_test.go
package main

import "testing"

func Test_firstStar(t *testing.T) {
	type args struct {
		input string
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{name: "abcdef", args: args{
			input: "2 3 0 3 10 11 12 1 1 0 1 99 2 1 1 2",
		}, want: 138},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := firstStar(tt.args.input); got != tt.want {
				t.Errorf("findOverlaps() = %v, want %v", got, tt.want)
			}
		})
	}
}
func Test_secondStar(t *testing.T) {
	type args struct {
		input string
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{name: "abcdef", args: args{
			input: "2 3 0 3 10 11 12 1 1 0 1 99 2 1 1 2",
		}, want: 66},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := secondStar(tt.args.input); got != tt.want {
				t.Errorf("findOverlaps() = %v, want %v", got, tt.want)
			}
		})
	}
}
